Enable IPv6 in SNMP
Strained at a camel and delivered a gnat. There may have been a
way to do this in local.conf, but I was impatient to fix the
problem. This leaves snmpd complaining about the lack of PCI,
but it keeps on going, as ifmob.patch for bug 2449 was alreay
in poky.